# schemas/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Image Studio Local API",
    "version": "1.0.0",
    "description": "OpenAPI contract for health, versioned jobs, image processing, and AI tools."
  },
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Local API health and engine capabilities",
        "responses": {
          "200": {
            "description": "Health payload with structured engines and capability flags",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs": {
      "post": {
        "operationId": "createJob",
        "summary": "Create an async job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/JobCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobStatus" }
              }
            }
          },
          "default": {
            "description": "Problem details",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetail" }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "getJob",
        "summary": "Get job status and progress",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobStatus" }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetail" }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancelJob",
        "summary": "Cancel a queued or running job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated job status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobStatus" }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetail" }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/result": {
      "get": {
        "operationId": "getJobResult",
        "summary": "Fetch job result when succeeded",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job result payload",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "404": {
            "description": "Job not found or result unavailable",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetail" }
              }
            }
          },
          "409": {
            "description": "Job not finished",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetail" }
              }
            }
          }
        }
      }
    },
    "/api/optimize-png": {
      "post": {
        "operationId": "optimizePng",
        "summary": "Optimize a PNG with the local Python backend",
        "responses": {
          "200": {
            "description": "Optimized PNG bytes",
            "content": {
              "image/png": { "schema": { "type": "string", "format": "binary" } }
            }
          }
        }
      }
    },
    "/api/ai/matte": {
      "post": {
        "operationId": "aiMatte",
        "summary": "Soft matte / background removal",
        "responses": {
          "200": {
            "description": "Matte result JSON",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
    "/api/ai/inpaint": {
      "post": {
        "operationId": "aiInpaint",
        "summary": "LaMa erase inpaint (OpenCV fallback)",
        "responses": {
          "200": {
            "description": "Inpainted image as base64 PNG JSON",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
    "/api/ai/detect": {
      "post": {
        "operationId": "aiDetect",
        "summary": "Backend prompt selection",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": { "type": "string", "format": "binary" },
                  "prompt": { "type": "string", "default": "" },
                  "confidence": { "type": "number", "default": 0.35 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detection result",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
    "/api/ai/upscale": {
      "post": {
        "operationId": "aiUpscale",
        "summary": "Upscale",
        "responses": {
          "200": {
            "description": "Upscale result",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "JobCreate": {
        "type": "object",
        "required": ["kind"],
        "properties": {
          "kind": { "type": "string" },
          "params": { "type": "object", "additionalProperties": true },
          "idempotency_key": { "type": "string", "nullable": true }
        }
      },
      "JobStatus": {
        "type": "object",
        "required": ["job_id", "status"],
        "properties": {
          "job_id": { "type": "string" },
          "kind": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["queued", "running", "succeeded", "failed", "cancelled"]
          },
          "progress": { "type": "number", "minimum": 0, "maximum": 1 },
          "error": { "type": "string", "nullable": true },
          "created_at": { "type": "number" },
          "updated_at": { "type": "number" }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "code": { "type": "string" },
          "request_id": { "type": "string" },
          "retryable": { "type": "boolean" }
        }
      }
    }
  }
}